Module Kv_intf_v3.Pvt_class_types

Generic syncable object

class type 'a mrshlr = object ... end

class type 'a set_once = object ... end

class type 'a mutable_ = object ... end

class type 'a on_disk_block = object ... end

This is a container for a type 'a that can be stored in a single block. The initialized method checks that all the state parts of a component are initialized; use with assert

class type 'fl freelist = object ... end

type t1 = {

mutable bt_rt : Tjr_fs_shared.Std_types.blk_id;

mutable pc_hd : Tjr_fs_shared.Std_types.blk_id;

mutable pc_tl : Tjr_fs_shared.Std_types.blk_id;

}

val bin_shape_t1 : Bin_prot.Shape.t

val bin_size_t1 : t1 Bin_prot.Size.sizer

val bin_write_t1 : t1 Bin_prot.Write.writer

val bin_writer_t1 : t1 Bin_prot.Type_class.writer

val __bin_read_t1__ : (int -> t1) Bin_prot.Read.reader

val bin_read_t1 : t1 Bin_prot.Read.reader

val bin_reader_t1 : t1 Bin_prot.Type_class.reader

val bin_t1 : t1 Bin_prot.Type_class.t

class type root_man = object ... end

class type btree = object ... end

class type 'kvop_map pcache = object ... end

class type ['k, 'v, 'mt_state] lru = object ... end

class type ['k, 'v, 'fl, 'kvop_map, 'mt_state] kv_store = object ... end

A collection of all the stateful components