Class type Pvt_class_types.on_disk_block

This is a container for a type 'a that can be stored in a single block. The initialized method checks that all the state parts of a component are initialized; use with assert

val blk_dev_ops : (Tjr_fs_shared.Std_types.blk_id, Tjr_fs_shared.Std_types.blk, Tjr_fs_shared.Std_types.t) Tjr_fs_shared.blk_dev_ops set_once

val blk_id : Tjr_fs_shared.Std_types.blk_id set_once

val contents : 'a mutable_

val marshaller : 'a mrshlr set_once

method sync_to_disk : unit -> (unit, Tjr_fs_shared.Std_types.t) Tjr_monad.m

method sync_from_disk : unit -> (unit, Tjr_fs_shared.Std_types.t) Tjr_monad.m

method is_initialized : bool