Module Tjr_kv__.Dmap_with_blocks_limit

type ('k, 'v, 't) dmap_with_lim_ops = ('k, 'v, 't) Tjr_fs_shared.Shared_map_ops.map_ops

val make_ops : monad_ops:'t Tjr_monad.monad_ops -> dmap_ops:('k, 'v, 'ptr, 'kvop_map, 't) Tjr_pcache.Pcache_intf.dmap_ops -> dmap_blocks_limit:int -> bt_find:('k -> ('v option, 't) Tjr_monad.m) -> bt_handle_detach:(('k, 'v, 'ptr, 'kvop_map) Tjr_pcache.Pcache_intf.detach_info -> (unit, 't) Tjr_monad.m) -> ('k, 'v, 't) Tjr_fs_shared.Shared_map_ops.map_ops

Construct the DMAP, which uses the dmap_ops and wraps it in a routine which occasionally executes a B-tree roll-up.

Parameters:

monad_ops

dmap_ops: dmap from tjr_pcache, with dmap interface

dmap_blocks_limit: how many blocks in the pcache before attempting a roll-up; if the length of pcache is >= this limit, we attempt a roll-up; NOTE that this limit should be >= 2 (if we roll up with 1 block, then in fact nothing gets rolled up because we roll up "upto" the current block; not a problem but probably pointless for testing)

bt_find: called if key not in pcache map FIXME do we need a write-through cache here? or just rely on the front-end LRU? FIXME note that even if a rollup is taking place, we can use the old B-tree root for the bt_find operation.

bt_handle_detach: called to detach the rollup into another thread; typically this operation puts a msg on a message queue which is then received and acted upon by the dedicated rollup thread