Module Tjr_kv__Kv_intf_v3

Interfaces based on a single "store" object with mutable refs.

NOTE the class types are not rendered well with odoc.

class type ['a] mrshlr = object
  method to_blk: 'a -> blk
  method of_blk: blk -> 'a
end

class type ['a] set_once = object
  method set: 'a -> unit
  method get: unit -> 'a
  method is_set: bool
end

class type ['a] mutable_ = object
  method set: 'a -> unit
  method get: unit -> 'a
  method is_set: bool
end

(** This is a container for a type 'a that can be stored in a single
   block. The initialized method checks that all the state parts of
   a component are initialized; use with assert *)
class type ['a] on_disk_block = object
  val blk_dev_ops    : (blk_id,blk,t)blk_dev_ops set_once
  val blk_id         : blk_id set_once
  val contents       : 'a mutable_
  val marshaller     : 'a mrshlr set_once
  method sync_to_disk   : unit -> (unit,t)m
  method sync_from_disk : unit -> (unit,t)m
  method is_initialized : bool
end

class type ['fl] freelist = object
  inherit ['fl] on_disk_block
  method blk_allocator: (r,t)blk_allocator_ops
end

type t1 = { 
  mutable bt_rt:blk_id;
  mutable pc_hd:blk_id; 
  mutable pc_tl:blk_id 
}[@@deriving bin_io]

class type root_man = object
  inherit [t1] on_disk_block
end

(* we expect that, once setup, the threads run forever; other
   applications may need to shutdown threads instead. *) 
class type btree = object
  method blk_dev_ops    : (blk_id,blk,t)blk_dev_ops set_once
  method blk_id         : blk_id mutable_ (* root of the btree *)
  method thread         : < start_thread: unit -> (unit,t)m >
  method btree_ops      : unit -> ('k,'v,blk_id,t) Kv_intf.Btree_ops.btree_ops
  method is_initialized : bool

  (** Flush the write back cache, after merging a prefix of the pcache *)
  method flush_cache    : unit -> (unit,t)m
end

class type ['kvop_map] pcache = object
  method blk_dev_ops    : (blk_id,blk,t)blk_dev_ops set_once
  method pcache_state   : (r,'kvop_map) Tjr_pcache.Pcache_state.pcache_state mutable_
  method thread         : < start_thread: unit -> (unit,t)m >
  method sync_to_disk   : unit -> (unit,t)m
  method sync_from_disk : hd:blk_id -> tl:blk_id -> (unit,t)m
  method is_initialized : bool
end

class type ['k,'v,'mt_state] lru = object
  method get_lru_state: 'mt_state mutable_
  method lru_ops: unit -> ('k,'v,t)mt_ops
end

(** A collection of all the stateful components *)
class type ['k,'v,'fl,'kvop_map,'mt_state] kv_store = object
  method root_man : root_man 
  method btree    : btree
  method q_pc_bt  : ('k,'v) Kv_intf_v2.q_pc_bt
  method pcache   : 'kvop_map pcache
  method q_lru_pc : ('k,'v) Kv_intf_v2.q_lru_pc
  method lru      : ('k,'v,'mt_state) lru

  (** Checks all components are initialized *)
  method is_initialized : bool
end

module Pvt_class_types : sig ... end

module Classes : sig ... end

include Classes

module P = Pvt_class_types

Generic syncable object

module Make_mrshlr = Classes.Make_mrshlr

class +'b mutable_ : object ... end

class +'b set_once : object ... end

class 'a on_disk_block : object ... end

val f : 'a on_disk_block -> 'a P.on_disk_block